+++ src/p5Drawer/debug.functions.js
@@ -3,14 +3,9 @@
 
 const Rac = require('../Rac');
 
-// RELEASE-TODO: update debug texts to use paddings instead of pointToAngle
 
-// RELEASE-TODO: use uprightText for these cases
-function reversesText(angle) {
-  return angle.turn < 3/4 && angle.turn >= 1/4;
-}
-
-
+// Creates and restores the drawing context for a dashed stroke while
+// `closure` is called.
 function dashedDraw(drawer, segment, closure) {
   const context = drawer.p5.drawingContext;
   context.save();
@@ -55,26 +50,20 @@ exports.debugAngle = function(drawer, angle, point, drawsText) {
   }
 
   // Debug Text
-  if (drawsText !== true) { return; }
+  if (drawsText !== true) return;
 
-  // Normal orientation
-  let format = new Rac.Text.Format(
-    rac,
+  let format = new Rac.Text.Format(rac,
     Rac.Text.Format.horizontalAlign.left,
     Rac.Text.Format.verticalAlign.center,
     angle,
     drawer.debugTextOptions.font,
-    drawer.debugTextOptions.size);
-  if (reversesText(angle)) {
-    // Reverse orientation
-    format = format.reverse();
-  }
+    drawer.debugTextOptions.size,
+    markerRadius*2, 0);
 
   // Turn text
   let turnString = `turn:${angle.turn.toFixed(digits)}`;
-  point
-    .pointToAngle(angle, markerRadius*2)
-    .text(turnString, format)
+  point.text(turnString, format)
+    .upright()
     .draw(drawer.debugTextStyle);
 }; // debugAngle
 
@@ -103,21 +92,50 @@ exports.debugPoint = function(drawer, point, drawsText) {
     .draw();
 
   // Debug Text
-  if (drawsText !== true) { return; }
+  if (drawsText !== true) return;
 
-  let string = `x:${point.x.toFixed(digits)}\ny:${point.y.toFixed(digits)}`;
   let format = new Rac.Text.Format(
     rac,
     Rac.Text.Format.horizontalAlign.left,
     Rac.Text.Format.verticalAlign.top,
     rac.Angle.e,
     drawer.debugTextOptions.font,
-    drawer.debugTextOptions.size);
-  point
-    .pointToAngle(rac.Angle.se, pointRadius*2)
-    .text(string, format)
+    drawer.debugTextOptions.size,
+    pointRadius, pointRadius);
+
+  let string = `x:${point.x.toFixed(digits)}\ny:${point.y.toFixed(digits)}`;
+  point.text(string, format)
     .draw(drawer.debugTextStyle);
 }; // debugPoint
+
+
+// Shared text drawing for ray and segment
+function drawRayTexts(drawer, ray, topString, bottomString) {
+  const hEnum = Rac.Text.Format.horizontalAlign;
+  const vEnum = Rac.Text.Format.verticalAlign;
+  const font        = drawer.debugTextOptions.font;
+  const size        = drawer.debugTextOptions.size;
+  const pointRadius = drawer.debugPointRadius;
+
+  let topFormat = new Rac.Text.Format(
+    drawer.rac,
+    hEnum.left, vEnum.bottom,
+    ray.angle, font, size,
+    pointRadius, pointRadius);
+  let bottomFormat = new Rac.Text.Format(
+    drawer.rac,
+    hEnum.left, vEnum.top,
+    ray.angle, font, size,
+    pointRadius, pointRadius);
+
+  // Texts
+  ray.text(topString, topFormat)
+    .upright()
+    .draw(drawer.debugTextStyle);
+  ray.text(bottomString, bottomFormat)
+    .upright()
+    .draw(drawer.debugTextStyle);
+};
 
 
 exports.debugRay = function(drawer, ray, drawsText) {
@@ -165,43 +183,12 @@ exports.debugRay = function(drawer, ray, drawsText) {
   }
 
   // Debug Text
-  if (drawsText !== true) { return; }
+  if (drawsText !== true) return;
 
-  const angle  = ray.angle;
-  const hEnum = Rac.Text.Format.horizontalAlign;
-  const vEnum = Rac.Text.Format.verticalAlign;
-  const font   = drawer.debugTextOptions.font;
-  const size   = drawer.debugTextOptions.size;
   const digits = drawer.debugTextOptions.fixedDigits;
-
-  // Normal orientation
-  let startFormat = new Rac.Text.Format(rac,
-    hEnum.left, vEnum.bottom,
-    angle, font, size);
-  let angleFormat = new Rac.Text.Format(rac,
-    hEnum.left, vEnum.top,
-    angle, font, size);
-  if (reversesText(angle)) {
-    // Reverse orientation
-    startFormat = startFormat.reverse();
-    angleFormat = angleFormat.reverse();
-  }
-
-  // Start text
   const startString = `start:(${ray.start.x.toFixed(digits)},${ray.start.y.toFixed(digits)})`;
-  ray.start
-    .pointToAngle(angle, pointRadius)
-    .pointToAngle(angle.subtract(1/4), markerRadius/2)
-    .text(startString, startFormat)
-    .draw(drawer.debugTextStyle);
-
-  // Angle text
-  const angleString = `angle:${angle.turn.toFixed(digits)}`;
-  ray.start
-    .pointToAngle(angle, pointRadius)
-    .pointToAngle(angle.add(1/4), markerRadius/2)
-    .text(angleString, angleFormat)
-    .draw(drawer.debugTextStyle);
+  const angleString = `angle:${ray.angle.turn.toFixed(digits)}`;
+  drawRayTexts(drawer, ray, startString, angleString);
 }; // debugRay
 
 
@@ -209,7 +196,6 @@ exports.debugSegment = function(drawer, segment, drawsText) {
   const rac =          drawer.rac;
   const pointRadius =  drawer.debugPointRadius;
   const markerRadius = drawer.debugMarkerRadius;
-  const digits =       drawer.debugTextOptions.fixedDigits;
 
   segment.draw();
 
@@ -264,47 +250,13 @@ exports.debugSegment = function(drawer, segment, drawsText) {
     .nextSegmentToPoint(endMarkerEnd.endPoint())
     .draw();
 
-
   // Debug Text
-  if (drawsText !== true) { return; }
+  if (drawsText !== true) return;
 
-  let angle = segment.angle();
-  // Normal orientation
-  let lengthFormat = new Rac.Text.Format(
-    rac,
-    Rac.Text.Format.horizontalAlign.left,
-    Rac.Text.Format.verticalAlign.bottom,
-    angle,
-    drawer.debugTextOptions.font,
-    drawer.debugTextOptions.size);
-  let angleFormat = new Rac.Text.Format(
-    rac,
-    Rac.Text.Format.horizontalAlign.left,
-    Rac.Text.Format.verticalAlign.top,
-    angle,
-    drawer.debugTextOptions.font,
-    drawer.debugTextOptions.size);
-  if (reversesText(angle)) {
-    // Reverse orientation
-    lengthFormat = lengthFormat.reverse();
-    angleFormat = angleFormat.reverse();
-  }
-
-  // Length
+  const digits = drawer.debugTextOptions.fixedDigits;
   let lengthString = `length:${segment.length.toFixed(digits)}`;
-  segment.startPoint()
-    .pointToAngle(angle, pointRadius)
-    .pointToAngle(angle.subtract(1/4), markerRadius/2)
-    .text(lengthString, lengthFormat)
-    .draw(drawer.debugTextStyle);
-
-    // Angle
-  let angleString = `angle:${angle.turn.toFixed(digits)}`;
-  segment.startPoint()
-    .pointToAngle(angle, pointRadius)
-    .pointToAngle(angle.add(1/4), markerRadius/2)
-    .text(angleString, angleFormat)
-    .draw(drawer.debugTextStyle);
+  let angleString  = `angle:${segment.ray.angle.turn.toFixed(digits)}`;
+  drawRayTexts(drawer, segment.ray, lengthString, angleString);
 }; // debugSegment
 
 
@@ -312,7 +264,6 @@ exports.debugArc = function(drawer, arc, drawsText) {
   const rac =          drawer.rac;
   const pointRadius =  drawer.debugPointRadius;
   const markerRadius = drawer.debugMarkerRadius;
-  const digits =       drawer.debugTextOptions.fixedDigits;
 
   arc.draw();
 
@@ -413,10 +364,13 @@ exports.debugArc = function(drawer, arc, drawsText) {
   }
 
   // Debug Text
-  if (drawsText !== true) { return; }
+  if (drawsText !== true) return;
 
   const hEnum = Rac.Text.Format.horizontalAlign;
   const vEnum = Rac.Text.Format.verticalAlign;
+  const font   = drawer.debugTextOptions.font;
+  const size   = drawer.debugTextOptions.size;
+  const digits = drawer.debugTextOptions.fixedDigits;
 
   let headVertical = arc.clockwise
     ? vEnum.top
@@ -428,41 +382,25 @@ exports.debugArc = function(drawer, arc, drawsText) {
     ? vEnum.bottom
     : vEnum.top;
 
-  // Normal orientation
-  let headFormat = new Rac.Text.Format(
-    rac,
-    hEnum.left,
-    headVertical,
+  let headFormat = new Rac.Text.Format(rac,
+    hEnum.left, headVertical,
     arc.start,
-    drawer.debugTextOptions.font,
-    drawer.debugTextOptions.size);
-  let tailFormat = new Rac.Text.Format(
-    rac,
-    hEnum.left,
-    tailVertical,
+    font, size,
+    pointRadius, 0);
+  let tailFormat = new Rac.Text.Format(rac,
+    hEnum.left, tailVertical,
     arc.end,
-    drawer.debugTextOptions.font,
-    drawer.debugTextOptions.size);
-  let radiusFormat = new Rac.Text.Format(
-    rac,
-    hEnum.left,
-    radiusVertical,
+    font, size,
+    pointRadius, 0);
+  let radiusFormat = new Rac.Text.Format(rac,
+    hEnum.left, radiusVertical,
     arc.start,
-    drawer.debugTextOptions.font,
-    drawer.debugTextOptions.size);
+    font, size,
+    markerRadius, pointRadius);
 
-  // Reverse orientation
-  if (reversesText(arc.start)) {
-    headFormat = headFormat.reverse();
-    radiusFormat = radiusFormat.reverse();
-  }
-  if (reversesText(arc.end)) {
-    tailFormat = tailFormat.reverse();
-  }
-
-  let startString = `start:${arc.start.turn.toFixed(digits)}`;
+  let startString  = `start:${arc.start.turn.toFixed(digits)}`;
   let radiusString = `radius:${arc.radius.toFixed(digits)}`;
-  let endString = `end:${arc.end.turn.toFixed(digits)}`;
+  let endString    = `end:${arc.end.turn.toFixed(digits)}`;
 
   let angleDistance = arc.angleDistance();
   let distanceString = `distance:${angleDistance.turn.toFixed(digits)}`;
@@ -471,36 +409,36 @@ exports.debugArc = function(drawer, arc, drawsText) {
   let headString;
 
   // Radius label
-  if (angleDistance.turn <= 3/4 && !arc.isCircle()) {
+  const endIsAway = angleDistance.turn <= 3/4 || angleDistance.equals(3/4);
+  if (endIsAway && !arc.isCircle()) {
     // Radius drawn separately
-    let perpAngle = arc.start.perpendicular(!arc.clockwise);
-    arc.center
-      .pointToAngle(arc.start, markerRadius)
-      .pointToAngle(perpAngle, pointRadius*2)
-      .text(radiusString, radiusFormat)
-      .draw(drawer.debugTextStyle);
     headString = startString;
+    arc.center
+      .text(radiusString, radiusFormat)
+      .upright()
+      .draw(drawer.debugTextStyle);
+
   } else {
     // Radius joined to head
     headString = `${startString}\n${radiusString}`;
   }
 
   if (lengthAtOrientationArc > textJoinThreshold) {
-    // Draw strings separately
+    // Draw head and tail separately
     orientationArc.startPoint()
-      .pointToAngle(arc.start, markerRadius/2)
       .text(headString, headFormat)
+      .upright()
       .draw(drawer.debugTextStyle);
     orientationArc.pointAtAngle(arc.end)
-      .pointToAngle(arc.end, markerRadius/2)
       .text(tailString, tailFormat)
+      .upright()
       .draw(drawer.debugTextStyle);
   } else {
-    // Draw strings together
-    let allStrings = `${headString}\n${tailString}`;
+    // Draw head and tail together
+    let bothStrings = `${headString}\n${tailString}`;
     orientationArc.startPoint()
-      .pointToAngle(arc.start, markerRadius/2)
-      .text(allStrings, headFormat)
+      .text(bothStrings, headFormat)
+      .upright()
       .draw(drawer.debugTextStyle);
   }
 }; // debugArc